refactor: break up ConversationEventReceiverTest
