+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/receiver/conversation/DeletedConversationEventHandlerTest.kt
@@ -1,0 +1,118 @@
+package com.wire.kalium.logic.sync.receiver.conversation
+
+import com.wire.kalium.logic.data.conversation.Conversation
+import com.wire.kalium.logic.data.conversation.ConversationRepository
+import com.wire.kalium.logic.data.id.ConversationId
+import com.wire.kalium.logic.data.user.UserId
+import com.wire.kalium.logic.data.user.UserRepository
+import com.wire.kalium.logic.feature.message.EphemeralNotificationsMgr
+import com.wire.kalium.logic.framework.TestConversation
+import com.wire.kalium.logic.framework.TestEvent
+import com.wire.kalium.logic.framework.TestUser
+import com.wire.kalium.logic.functional.Either
+import io.mockative.Mock
+import io.mockative.any
+import io.mockative.classOf
+import io.mockative.eq
+import io.mockative.given
+import io.mockative.mock
+import io.mockative.once
+import io.mockative.thenDoNothing
+import io.mockative.verify
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.test.runTest
+import kotlin.test.Test
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class DeletedConversationEventHandlerTest {
+
+    @Test
+    fun givenADeletedConversationEvent_whenHandlingItAndNotExists_thenShouldSkipTheDeletion() = runTest {
+        val event = TestEvent.deletedConversation()
+        val (arrangement, eventHandler) = Arrangement()
+            .withEphemeralNotificationEnqueue()
+            .withGetConversation(null)
+            .withGetUserAuthor(event.senderUserId)
+            .withDeletingConversationSucceeding()
+            .arrange()
+
+        eventHandler.handle(event)
+
+        with(arrangement) {
+            verify(conversationRepository)
+                .suspendFunction(conversationRepository::deleteConversation)
+                .with(eq(TestConversation.ID))
+                .wasNotInvoked()
+        }
+    }
+
+    @Test
+    fun givenADeletedConversationEvent_whenHandlingIt_thenShouldDeleteTheConversationAndItsContent() = runTest {
+        val event = TestEvent.deletedConversation()
+        val (arrangement, eventHandler) = Arrangement()
+            .withEphemeralNotificationEnqueue()
+            .withGetConversation()
+            .withGetUserAuthor(event.senderUserId)
+            .withDeletingConversationSucceeding()
+            .arrange()
+
+        eventHandler.handle(event)
+
+        with(arrangement) {
+            verify(conversationRepository)
+                .suspendFunction(conversationRepository::deleteConversation)
+                .with(eq(TestConversation.ID))
+                .wasInvoked(exactly = once)
+        }
+    }
+
+    private class Arrangement {
+
+        @Mock
+        val conversationRepository = mock(classOf<ConversationRepository>())
+
+        @Mock
+        private val userRepository = mock(classOf<UserRepository>())
+
+        @Mock
+        private val ephemeralNotifications = mock(classOf<EphemeralNotificationsMgr>())
+
+        private val deletedConversationEventHandler: DeletedConversationEventHandler = DeletedConversationEventHandlerImpl(
+            userRepository,
+            conversationRepository,
+            ephemeralNotifications
+        )
+
+        fun withDeletingConversationSucceeding(conversationId: ConversationId = TestConversation.ID) = apply {
+            given(conversationRepository)
+                .suspendFunction(conversationRepository::deleteConversation)
+                .whenInvokedWith((eq(conversationId)))
+                .thenReturn(Either.Right(Unit))
+        }
+
+        fun withGetConversation(conversation: Conversation? = TestConversation.CONVERSATION) = apply {
+            given(conversationRepository)
+                .suspendFunction(conversationRepository::getConversationById)
+                .whenInvokedWith(any())
+                .thenReturn(conversation)
+        }
+
+        fun withGetUserAuthor(userId: UserId = TestUser.USER_ID) = apply {
+            given(userRepository)
+                .suspendFunction(userRepository::observeUser)
+                .whenInvokedWith(eq(userId))
+                .thenReturn(flowOf(TestUser.OTHER))
+        }
+
+        fun withEphemeralNotificationEnqueue() = apply {
+            given(ephemeralNotifications)
+                .suspendFunction(ephemeralNotifications::scheduleNotification)
+                .whenInvokedWith(any())
+                .thenDoNothing()
+        }
+
+        fun arrange() = this to deletedConversationEventHandler
+    }
+
+}
