+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/receiver/conversation/MemberJoinEventHandlerTest.kt
@@ -1,0 +1,166 @@
+package com.wire.kalium.logic.sync.receiver.conversation
+
+import com.wire.kalium.logic.CoreFailure
+import com.wire.kalium.logic.NetworkFailure
+import com.wire.kalium.logic.StorageFailure
+import com.wire.kalium.logic.data.conversation.Conversation.Member
+import com.wire.kalium.logic.data.conversation.ConversationRepository
+import com.wire.kalium.logic.data.message.Message
+import com.wire.kalium.logic.data.message.MessageContent
+import com.wire.kalium.logic.data.message.PersistMessageUseCase
+import com.wire.kalium.logic.data.user.UserRepository
+import com.wire.kalium.logic.framework.TestEvent
+import com.wire.kalium.logic.framework.TestUser
+import com.wire.kalium.logic.functional.Either
+import io.mockative.Mock
+import io.mockative.any
+import io.mockative.classOf
+import io.mockative.eq
+import io.mockative.given
+import io.mockative.matching
+import io.mockative.mock
+import io.mockative.once
+import io.mockative.verify
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runTest
+import kotlin.test.Test
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class MemberJoinEventHandlerTest {
+
+    @Test
+    fun givenMemberJoinEvent_whenHandlingIt_thenShouldFetchConversationIfUnknown() = runTest {
+        val newMembers = listOf(Member(TestUser.USER_ID, Member.Role.Member))
+        val event = TestEvent.memberJoin(members = newMembers)
+
+        val (arrangement, eventHandler) = Arrangement()
+            .withPersistingMessageReturning(Either.Right(Unit))
+            .withFetchConversationIfUnknownSucceeding()
+            .withPersistMembersSucceeding()
+            .withFetchUsersIfUnknownByIdsReturning(Either.Right(Unit))
+            .arrange()
+
+        eventHandler.handle(event)
+
+        verify(arrangement.conversationRepository)
+            .suspendFunction(arrangement.conversationRepository::fetchConversationIfUnknown)
+            .with(eq(event.conversationId))
+            .wasInvoked(exactly = once)
+    }
+
+    @Test
+    fun givenMemberJoinEvent_whenHandlingIt_thenShouldPersistMembers() = runTest {
+        val newMembers = listOf(Member(TestUser.USER_ID, Member.Role.Member))
+        val event = TestEvent.memberJoin(members = newMembers)
+
+        val (arrangement, eventHandler) = Arrangement()
+            .withPersistingMessageReturning(Either.Right(Unit))
+            .withFetchConversationIfUnknownSucceeding()
+            .withPersistMembersSucceeding()
+            .withFetchUsersIfUnknownByIdsReturning(Either.Right(Unit))
+            .arrange()
+
+        eventHandler.handle(event)
+
+        verify(arrangement.conversationRepository)
+            .suspendFunction(arrangement.conversationRepository::persistMembers)
+            .with(eq(newMembers), eq(event.conversationId))
+            .wasInvoked(exactly = once)
+    }
+
+    @Test
+    fun givenMemberJoinEventAndFetchConversationFails_whenHandlingIt_thenShouldAttemptPersistingMembersAnyway() = runTest {
+        val newMembers = listOf(Member(TestUser.USER_ID, Member.Role.Member))
+        val event = TestEvent.memberJoin(members = newMembers)
+
+        val (arrangement, eventHandler) = Arrangement()
+            .withPersistingMessageReturning(Either.Right(Unit))
+            .withFetchConversationIfUnknownFailing(NetworkFailure.NoNetworkConnection(null))
+            .withPersistMembersSucceeding()
+            .arrange()
+
+        eventHandler.handle(event)
+
+        verify(arrangement.conversationRepository)
+            .suspendFunction(arrangement.conversationRepository::persistMembers)
+            .with(eq(newMembers), eq(event.conversationId))
+            .wasInvoked(exactly = once)
+    }
+
+    @Test
+    fun givenMemberJoinEvent_whenHandlingIt_thenShouldPersistSystemMessage() = runTest {
+        val newMembers = listOf(Member(TestUser.USER_ID, Member.Role.Admin))
+        val event = TestEvent.memberJoin(members = newMembers)
+
+        val (arrangement, eventHandler) = Arrangement()
+            .withPersistingMessageReturning(Either.Right(Unit))
+            .withFetchConversationIfUnknownFailing(NetworkFailure.NoNetworkConnection(null))
+            .withPersistMembersSucceeding()
+            .arrange()
+
+        eventHandler.handle(event)
+
+        verify(arrangement.persistMessage)
+            .suspendFunction(arrangement.persistMessage::invoke)
+            .with(
+                matching {
+                    it is Message.System && it.content is MessageContent.MemberChange
+                }
+            )
+            .wasInvoked(exactly = once)
+    }
+
+    private class Arrangement {
+        @Mock
+        val persistMessage = mock(classOf<PersistMessageUseCase>())
+
+        @Mock
+        val conversationRepository = mock(classOf<ConversationRepository>())
+
+        @Mock
+        private val userRepository = mock(classOf<UserRepository>())
+
+        private val memberJoinEventHandler: MemberJoinEventHandler = MemberJoinEventHandlerImpl(
+            conversationRepository,
+            persistMessage
+        )
+
+        fun withPersistingMessageReturning(result: Either<CoreFailure, Unit>) = apply {
+            given(persistMessage)
+                .suspendFunction(persistMessage::invoke)
+                .whenInvokedWith(any())
+                .thenReturn(result)
+        }
+
+        fun withFetchConversationIfUnknownSucceeding() = apply {
+            given(conversationRepository)
+                .suspendFunction(conversationRepository::fetchConversationIfUnknown)
+                .whenInvokedWith(any())
+                .thenReturn(Either.Right(Unit))
+        }
+
+        fun withFetchConversationIfUnknownFailing(coreFailure: CoreFailure) = apply {
+            given(conversationRepository)
+                .suspendFunction(conversationRepository::fetchConversationIfUnknown)
+                .whenInvokedWith(any())
+                .thenReturn(Either.Left(coreFailure))
+        }
+
+        fun withPersistMembersSucceeding() = apply {
+            given(conversationRepository)
+                .suspendFunction(conversationRepository::persistMembers)
+                .whenInvokedWith(any(), any())
+                .thenReturn(Either.Right(Unit))
+        }
+
+        fun withFetchUsersIfUnknownByIdsReturning(result: Either<StorageFailure, Unit>) = apply {
+            given(userRepository)
+                .suspendFunction(userRepository::fetchUsersIfUnknownByIds)
+                .whenInvokedWith(any())
+                .thenReturn(result)
+        }
+
+        fun arrange() = this to memberJoinEventHandler
+    }
+
+}
